Reverse internal order of tabs, for symmetry
Vim tabs are asymmetric. They open on the right, but when you close a
tab, you don't return to the tab on the left.

This can be handled by calling tabprevious every time a tab is closed.
However some plugins, vim-fugitive for instance, avoid the asymmetry by
opening new tabs on the left.
So as to avoid issues, we instead reverse the order in which tabs are
displayed. This means new tabs opened in the default way then have to be
moved into the correct position, but no special handling of tab closing
is required. Now vim-fugitive works as originally intended.